+++ src/main/java/com/cburch/logisim/gui/prefs/PreferencesFrame.java
@@ -45,6 +45,7 @@ public class PreferencesFrame extends LFrame.Dialog {
           new SoftwaresOptions(this),
           new FpgaOptions(this),
           new HotkeyOptions(this),
+          new AutosaveOptions(this),
         };
     tabbedPane = new JTabbedPane();
     int intlIndex = -1;